requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9

# Data Processing (Core for SchemeProcessor)
pandas>=2.0.0
numpy>=1.24.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# REMOVED: Excel dependencies (no longer needed)
# openpyxl>=3.1.0
# xlrd>=2.0.1 
# XlsxWriter>=3.1.9

# REMOVED: Async database (using direct psycopg2 like astra-main)
# asyncpg==0.29.0
# sqlalchemy==2.0.23
# asyncio-throttle==1.0.2
# pydantic-settings==2.1.0

# NOTE: This is now optimized for SchemeProcessor from astra-main
# - No Excel file operations
# - No Streamlit dependencies
# - Direct database connectivity like astra-main
# - JSON-only responses